Reverted organization
Changed package organization back to default.
