+++ src/org/usfirst/frc/team840/robot/RobotMap.java
@@ -1,4 +1,4 @@
-package edu.art.frc.recycleRush.robot;
+package org.usfirst.frc.team840.robot;
 /**
  * The RobotMap is a mapping from the ports sensors and actuators are wired into
  * to a variable name. This provides flexibility changing wiring, makes checking
